+++ qs_getx_tool/test/support/test_dispose_controller.dart
@@ -1,0 +1,3 @@
+import 'package:qs_getx_tool/qs_getx_tool.dart';
+
+class TestDisposeController extends QsDisposeController {}