+++ qs_getx_tool/test/support/tracked_change_notifier.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/foundation.dart';
+
+class TrackedChangeNotifier extends ChangeNotifier {
+  @override
+  void dispose() {
+    isDisposed = true;
+    super.dispose();
+  }
+
+  bool isDisposed = false;
+}